rename manual-checkin to guest-list-checkin
